Documents and changes (a bit) Player class
Write documentation explaining what should be done.

While writing it I made a few changes. The idea was to change the match
history to a dictionary which saves (for asymmetric games) both the
last matches played against a given player as row/col player and as
either one.

Tests still need to be written for this!